I'm going to the cardiologist today. :)
I had a small attack of arrhythmia during the summer (my heart was
beating at 240 bpm :O ) and I went to the doctor's office. They did some
tests and took my blood (12 vials of it to be exact. >:(  I think my
nurse was secretly a vampire or a black market organ dealer). And now
for some reason I'm going to see a cardiologist. :)

P.S. This thing told me to download "Powershell" or something like that
in order to commit. I wonder if I'm going to turn into one of the turtle
things from TMNT. :D

P.S.S. My username is Hydrozoa